refactor(collaboration): extract chat lifecycle authority
